Rewrite constraint satisfaction.
Propagate requirements through declarations as expressions,
not constraints.
This touches the satisfaction rules, subtitution through
requirements, and diagnostics, and declaration analysis.
Stop normalizing constraints prior to evaluation; do this at
the same time.
Mostly disable the former diagnostic system, this will need
to be redesigned due to the changes to the simpler satisfaction
system.
Effectively disable terse notation for now.